feat(slugbuilder): run slugbuilder as non-root
This is a security fix as well as a fix for buildpacks that cannot
be run as root.

One small fix had to be made here with openssl. Docker by default
hardcodes $HOME to /, so shifting the export HOME line up before we
call openssl gives openssl write access to $HOME/.rnd.

fixes deis/deis#568

Signed-off-by: Matthew Fisher <matthewf@opdemand.com>